pyproject.toml
[project]
name = "agent-platform"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    # Core AI providers
    "anthropic>=0.40.0",
    "openai>=1.50.0",
    "google-generativeai>=0.8.0",
    "ollama>=0.4.0",
    
    # RAG stack
    "chromadb>=0.5.0",
    "pdfplumber>=0.11.0",
    "python-docx>=1.0.0",
    "tiktoken>=0.7.0",
    
    # Infrastructure
    "fastapi>=0.115.0",
    "uvicorn>=0.34.0",
    "pydantic>=2.10.0",
    "pyyaml>=6.0.1",
    "aiohttp>=3.11.0",
    "httpx>=0.28.0",
    "tenacity>=9.0.0",
    
    # Data & persistence
    "sqlalchemy>=2.0.20",
    "aiosqlite>=0.20.0",
    
    # MCP
    "mcp>=1.4.0",
    
    # CLI
    "typer>=0.15.0",
    "rich>=13.9.0",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "factory-boy>=3.3.0",
]
audio = [
    "whisper>=1.1.0",    # For audio transcript ingestion
    "torch>=2.0.0",
]

[project.scripts]
agent = "core.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["core", "rag", "memory", "skills", "workflows", "router", "mcp", "tracing"]
